+++ src/Attribute/Operation.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Nelmio\ApiDocBundle\Annotation;
+namespace Nelmio\ApiDocBundle\Attribute;
 
 use OpenApi\Annotations\Operation as BaseOperation;
