+++ tests/Functional/Controller/PromotedPropertiesController81.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Nelmio\ApiDocBundle\Tests\Functional\Controller;
 
-use Nelmio\ApiDocBundle\Annotation\Model;
+use Nelmio\ApiDocBundle\Attribute\Model;
 use Nelmio\ApiDocBundle\Tests\Functional\Entity\EntityWithPromotedPropertiesWithDefaults;
 use OpenApi\Attributes as OA;
 use Symfony\Component\Routing\Annotation\Route;
